+++ test/manifest.js
@@ -39,3 +39,4 @@ require('./test_inheritance');
 require('./test_and');
 require('./test_or');
 require('./test_or_else');
+require('./test_as_promise');